refactor(umbp): extract IRespClient seam so the store is client-agnostic
Split RespValue/RespError and a new IRespClient interface (Command/Eval/
EvalPipeline/Ping) into resp_value.h; RespClient now implements IRespClient, and
the store holds vector<unique_ptr<IRespClient>> instead of the concrete client.
No behavior change (pure refactor) — this lets a redis-plus-plus cluster client
implement the same seam for Redis Cluster mode while the hiredis client keeps
driving single / multi-endpoint. All 49 backend tests still pass.
